careerlens_api/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg[binary]
pydantic
python-dotenv
requests
pypdf
python-docx
